requirements.txt
flask
flask-cors
pymongo
python-dotenv
pandas
scikit-learn==1.6.0
xgboost